refactor: replace useTextSelectionPosition with useSelectionMenuOffset for improved text selection menu positioning
